add postAuthorize to controller

+++ core/java/iesi-rest-without-microservices/src/main/java/io/metadew/iesi/server/rest/component/ComponentsController.java
@@ -3,10 +3,12 @@ package io.metadew.iesi.server.rest.component;
 import io.metadew.iesi.metadata.configuration.exception.MetadataAlreadyExistsException;
 import io.metadew.iesi.metadata.configuration.exception.MetadataDoesNotExistException;
 import io.metadew.iesi.metadata.definition.component.key.ComponentKey;
+import io.metadew.iesi.metadata.service.user.IESIPrivilege;
 import io.metadew.iesi.metadata.tools.IdentifierTools;
 import io.metadew.iesi.server.rest.component.dto.ComponentDto;
 import io.metadew.iesi.server.rest.component.dto.ComponentDtoResourceAssembler;
 import io.metadew.iesi.server.rest.component.dto.IComponentDtoService;
+import io.metadew.iesi.server.rest.configuration.security.IesiSecurityChecker;
 import io.metadew.iesi.server.rest.error.DataBadRequestException;
 import io.metadew.iesi.server.rest.resource.HalMultipleEmbeddedResource;
 import org.springframework.beans.factory.annotation.Autowired;
@@ -17,6 +19,8 @@ import org.springframework.data.web.PagedResourcesAssembler;
 import org.springframework.hateoas.PagedModel;
 import org.springframework.http.HttpStatus;
 import org.springframework.http.ResponseEntity;
+import org.springframework.security.access.AccessDeniedException;
+import org.springframework.security.access.prepost.PostAuthorize;
 import org.springframework.security.access.prepost.PreAuthorize;
 import org.springframework.security.core.context.SecurityContextHolder;
 import org.springframework.web.bind.annotation.*;
@@ -38,16 +42,20 @@ public class ComponentsController {
     private final IComponentDtoService componentDtoService;
     private final ComponentDtoResourceAssembler componentDtoResourceAssembler;
     private final PagedResourcesAssembler<ComponentDto> componentDtoPagedResourcesAssembler;
+    private final IesiSecurityChecker iesiSecurityChecker;
 
     @Autowired
     ComponentsController(ComponentDtoResourceAssembler componentDtoResourceAssembler,
                          IComponentService componentService,
                          IComponentDtoService componentDtoService,
-                         PagedResourcesAssembler<ComponentDto> componentDtoPagedResourcesAssembler) {
+                         PagedResourcesAssembler<ComponentDto> componentDtoPagedResourcesAssembler,
+                         IesiSecurityChecker iesiSecurityChecker
+    ) {
         this.componentDtoResourceAssembler = componentDtoResourceAssembler;
         this.componentService = componentService;
         this.componentDtoService = componentDtoService;
         this.componentDtoPagedResourcesAssembler = componentDtoPagedResourcesAssembler;
+        this.iesiSecurityChecker = iesiSecurityChecker;
     }
 
     @GetMapping("")
@@ -82,6 +90,7 @@ public class ComponentsController {
 
     @GetMapping("/{name}/{version}")
     @PreAuthorize("hasPrivilege('COMPONENTS_READ')")
+    @PostAuthorize("hasPrivilege('COMPONENTS_READ', returnObject.securityGroupName)")
     public ComponentDto get(@PathVariable String name, @PathVariable Long version) throws MetadataDoesNotExistException {
         ComponentDto component = componentDtoService.getByNameAndVersion(SecurityContextHolder.getContext().getAuthentication(), name, version)
                 .orElseThrow(() -> new MetadataDoesNotExistException(new ComponentKey(IdentifierTools.getComponentIdentifier(name), version)));
@@ -89,7 +98,7 @@ public class ComponentsController {
     }
 
     @PostMapping("")
-    @PreAuthorize("hasPrivilege('COMPONENTS_WRITE')")
+    @PreAuthorize("hasPrivilege('COMPONENTS_WRITE', #componentDto.securityGroupName)")
     public ComponentDto post(@Valid @RequestBody ComponentDto componentDto) {
         try {
             componentService.createComponent(componentDto);
@@ -101,7 +110,7 @@ public class ComponentsController {
     }
 
     @PutMapping("")
-    @PreAuthorize("hasPrivilege('COMPONENTS_WRITE')")
+    @PreAuthorize("hasPrivilege('COMPONENTS_WRITE', #componentDtos.![securityGroupName])")
     public HalMultipleEmbeddedResource<ComponentDto> putAll(@Valid @RequestBody List<ComponentDto> componentDtos) throws MetadataDoesNotExistException {
         componentService.updateComponents(componentDtos);
         HalMultipleEmbeddedResource<ComponentDto> halMultipleEmbeddedResource = new HalMultipleEmbeddedResource<>();
@@ -116,7 +125,7 @@ public class ComponentsController {
     }
 
     @PutMapping("/{name}/{version}")
-    @PreAuthorize("hasPrivilege('COMPONENTS_WRITE')")
+    @PreAuthorize("hasPrivilege('COMPONENTS_WRITE', #componentDto.securityGroupName)")
     public ComponentDto put(@PathVariable String name, @PathVariable Long version, @RequestBody ComponentDto componentDto) throws MetadataDoesNotExistException {
         if (!componentDto.getName().equals(name)) {
             throw new DataBadRequestException(name);
@@ -128,24 +137,19 @@ public class ComponentsController {
 
     }
 
-    @DeleteMapping("")
-    @PreAuthorize("hasPrivilege('COMPONENTS_WRITE')")
-    public ResponseEntity<?> deleteAll() {
-        componentService.deleteAll();
-        return ResponseEntity.status(HttpStatus.OK).build();
-    }
-
-    @DeleteMapping("/{name}")
-    @PreAuthorize("hasPrivilege('COMPONENTS_WRITE')")
-    public ResponseEntity<?> deleteByName(@PathVariable String name) {
-        componentService.deleteByName(name);
-        return ResponseEntity.status(HttpStatus.OK).build();
-    }
-
     @DeleteMapping("/{name}/{version}")
     @PreAuthorize("hasPrivilege('COMPONENTS_WRITE')")
     public ResponseEntity<?> delete(@PathVariable String name, @PathVariable Long version) throws MetadataDoesNotExistException {
+        ComponentDto componentDto = componentDtoService.getByNameAndVersion(null, name, version)
+                .orElseThrow(() -> new MetadataDoesNotExistException(new ComponentKey(IdentifierTools.getComponentIdentifier(name), version)));
+        if (!iesiSecurityChecker.hasPrivilege(SecurityContextHolder.getContext().getAuthentication(),
+                IESIPrivilege.SCRIPTS_MODIFY.getPrivilege(),
+                componentDto.getSecurityGroupName())
+        ) {
+            throw new AccessDeniedException("User is not allowed to delete this component");
+        }
         componentService.deleteByNameAndVersion(name, version);
+
         return ResponseEntity.status(HttpStatus.OK).build();
     }
 }
